graph_lab_sql/requirements.txt
asyncpg>=0.29
pgvector>=0.3
numpy>=1.26
pydantic>=2.0
structlog>=24.0
pyyaml>=6.0
jinja2>=3.1
python-dotenv>=1.0
anthropic>=0.40
openai>=1.50
httpx>=0.27
tiktoken>=0.7
typer>=0.12
rich>=13.0
pytest>=8.0
pytest-asyncio>=0.24
